# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: Mini Blog API
  version: 1.0.0
  description: API REST para gestionar autores y posts de un mini blog.
servers:
  - url: http://localhost:3000
    description: Servidor local
...
paths:
  /api/authors:
    get:
      tags:
        - Authors
      summary: Listar todos los autores
      responses:
        '200':
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'

    post:
      tags:
        - Authors
      summary: Crear un nuevo autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthor'
      responses:
        '201':
          description: Autor creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/posts:
    get:
      tags:
        - Posts
      summary: Listar todos los posts
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostWithAuthor'

    post:
      tags:
        - Posts
      summary: Crear un nuevo post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '201':
          description: Post creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Ana López
        email:
          type: string
          example: ana@example.com
        bio:
          type:
            - string
            - "null"
          example: Escritora de tecnología
        created_at:
          type: string
          format: date-time
          example: "2025-03-01T10:00:00Z"
    CreateAuthor:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: Ana López
        email:
          type: string
          example: ana@example.com
        bio:
          type:
            - string
            - "null"
          example: Escritora de tecnología
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Mi primer post
        content:
          type: string
          example: Este es el contenido del post.
        author_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2025-03-01T10:30:00Z"
    CreatePost:
      type: object
      required:
        - title
        - content
        - author_id
      properties:
        title:
          type: string
          example: Mi primer post
        content:
          type: string
          example: Este es el contenido del post.
        author_id:
          type: integer
          example: 1
    PostWithAuthor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Mi primer post
        content:
          type: string
          example: Este es el contenido del post.
        author_id:
          type: integer
          example: 1
        author_name:
          type: string
          example: Ana López
        author_email:
          type: string
          example: ana@example.com
        created_at:
          type: string
          format: date-time
          example: "2025-03-01T10:30:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: Recurso no encontrado